[BOOK-67] refactor: 컴포넌트 스캔을 위한 패키지이동
