+++ admin/src/main/kotlin/org/yapp/admin/config/InfraConfig.kt
@@ -1,8 +1,8 @@
 package org.yapp.admin.config
 
 import org.springframework.context.annotation.Configuration
-import org.yapp.EnableInfraBaseConfig
-import org.yapp.InfraBaseConfigGroup
+import org.yapp.infra.EnableInfraBaseConfig
+import org.yapp.infra.InfraBaseConfigGroup
 
 @Configuration(proxyBeanMethods = false)
 @EnableInfraBaseConfig([InfraBaseConfigGroup.JPA])
