+++ infra/src/main/kotlin/org/yapp/infra/InfraBaseConfig.kt
@@ -1,4 +1,4 @@
-package org.yapp
+package org.yapp.infra
 
 interface InfraBaseConfig {
 }
