+++ infra/src/main/kotlin/org/yapp/infra/InfraBaseConfigGroup.kt
@@ -1,8 +1,8 @@
-package org.yapp
+package org.yapp.infra
 
-import org.yapp.config.internal.async.AsyncConfig
-import org.yapp.config.internal.jpa.JpaConfig
-import org.yapp.config.external.redis.RedisConfig
+import org.yapp.infra.config.internal.async.AsyncConfig
+import org.yapp.infra.config.internal.jpa.JpaConfig
+import org.yapp.infra.config.external.redis.RedisConfig
 
 enum class InfraBaseConfigGroup(
     val configClass: Class<out InfraBaseConfig>
