+++ infra/src/main/kotlin/org/yapp/infra/config/external/redis/RedisConfig.kt
@@ -1,21 +1,19 @@
-package org.yapp.config.external.redis
+package org.yapp.infra.config.external.redis
 
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.ComponentScan
 import org.springframework.context.annotation.Configuration
 import org.springframework.data.redis.connection.RedisConnectionFactory
 import org.springframework.data.redis.connection.RedisStandaloneConfiguration
 import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactory
 import org.springframework.data.redis.core.RedisTemplate
 import org.springframework.data.redis.serializer.StringRedisSerializer
-import org.yapp.InfraBaseConfig
+import org.yapp.infra.InfraBaseConfig
 
 /**
  * Configuration for Redis.
  */
 @Configuration
-@ComponentScan("org.yapp.redis")
 class RedisConfig : InfraBaseConfig {
 
     @Value("\${spring.data.redis.host:localhost}")
