+++ infra/src/main/kotlin/org/yapp/infra/config/internal/jpa/JpaConfig.kt
@@ -1,9 +1,9 @@
-package org.yapp.config.internal.jpa
+package org.yapp.infra.config.internal.jpa
 
 import org.springframework.boot.autoconfigure.domain.EntityScan
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories
 import org.springframework.transaction.annotation.EnableTransactionManagement
-import org.yapp.InfraBaseConfig
+import org.yapp.infra.InfraBaseConfig
 
 @EnableTransactionManagement
 @EntityScan(basePackages = ["org.yapp"])
